packages/cra-template-typescript/template/scripts/requirements.txt
chardet==3.0.4
urllib3==1.25.7
requests==2.22.0
boto3==1.10.45
GitPython==3.1.7
PyGithub==1.53
